requirements.txt
arabic-reshaper==3.0.0
arrow==1.3.0
asgiref==3.9.1
asn1crypto==1.5.1
binaryornot==0.4.4
Brotli==1.1.0
certifi==2025.7.9
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
cloudinary==1.44.1
cookiecutter==2.6.0
cryptography==45.0.5
cssselect2==0.8.0
defusedxml==0.7.1
dj-database-url==3.0.1
Django==5.2.4
django-allauth==65.10.0
django-appconf==1.1.0
django-browser-reload==1.18.0
django-cloudinary-storage==0.3.0
django-colorfield==0.14.0
django-cotton==2.1.3
django-dynamic-breadcrumbs==0.4.3
django-extensions==4.1
django-filter==25.1
django-htmx==1.23.2
django-image-cropping==1.7
django-shortcuts==1.6
django-sslserver==0.22
django-tailwind==4.2.0
django-view-breadcrumbs==2.5.1
easy-thumbnails==2.10
fonttools==4.59.0
html5lib==1.1
idna==3.10
Jinja2==3.1.6
lxml==6.0.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
oauthlib==3.3.1
oscrypto==1.3.0
pillow==11.3.0
psycopg==3.2.9
pycparser==2.22
pydyf==0.11.0
Pygments==2.19.2
pyHanko==0.29.1
pyhanko-certvalidator==0.27.0
PyJWT==2.10.1
pypdf==5.9.0
pyphen==0.17.2
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-slugify==8.0.4
python3-openid==3.2.0
PyYAML==6.0.2
razorpay==1.4.2
reportlab==4.4.3
requests==2.32.4
requests-oauthlib==2.0.0
rich==14.0.0
setuptools==80.9.0
six==1.17.0
social-auth-core==4.7.0
sqlparse==0.5.3
svglib==1.5.1
text-unidecode==1.3
tinycss2==1.4.0
tinyhtml5==2.0.0
types-python-dateutil==2.9.0.20250708
typing_extensions==4.14.1
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.5.0
weasyprint==66.0
webencodings==0.5.1
xhtml2pdf==0.2.17
zopfli==0.2.3.post1
